+++ plugins/Pluginer.php
@@ -42,7 +42,12 @@ class Pluginer extends Admin
 
 	// applyPlugin
 
-	public function name()
+	public function init(): void
+	{
+		$this->applyPlugin(__FUNCTION__, func_get_args());
+	}
+
+	public function getServiceTitle(): string
 	{
 		return $this->applyPlugin(__FUNCTION__, func_get_args());
 	}
@@ -92,7 +97,7 @@ class Pluginer extends Admin
 		return $this->applyPlugin(__FUNCTION__, func_get_args());
 	}
 
-	public function queryTimeout()
+	public function getQueryTimeout(): int
 	{
 		return $this->applyPlugin(__FUNCTION__, func_get_args());
 	}
@@ -128,6 +133,11 @@ class Pluginer extends Admin
 	public function getLoginFormRow(string $fieldName, string $label, string $field): string
 	{
 		return $this->applyPlugin(__FUNCTION__, func_get_args());
+	}
+
+	public function printLogout(): void
+	{
+		$this->applyPlugin(__FUNCTION__, func_get_args());
 	}
 
 	public function getTableName(array $tableStatus): string